Create a separate script for the hub app
