+++ pyproject.toml
@@ -41,6 +41,7 @@ dev = ["ruff==0.4.4"]
 
 [project.scripts]
 jupyterlab-gallery = "jupyterlab_gallery:GalleryApp.launch_instance"
+jupyterhub-gallery = "jupyterlab_gallery.hub:HubGalleryApp.launch_instance"
 
 [tool.hatch.version]
 source = "nodejs"
